// apps/web/src/services/generated/models/createSystemRoleDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * API Documentation
 *
 *     当前环境: development
 *     最后更新: 9/15/2026, 5:29:26 PM
 *     [项目设计文档](https://wiki.example.com)
 *
 * OpenAPI spec version: 1.0
 */

export interface CreateSystemRoleDto {
  /**
   * 系统角色标识
   * @maxLength 64
   */
  key: string;
  /** 角色显示名称 */
  label: string;
  /**
   * 角色说明
   * @maxLength 1024
   */
  description?: string;
}