// apps/web/src/services/generated/models/systemRoleVo.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * API Documentation
 *
 *     当前环境: development
 *     最后更新: 9/15/2026, 5:29:26 PM
 *     [项目设计文档](https://wiki.example.com)
 *
 * OpenAPI spec version: 1.0
 */

export interface SystemRoleVo {
  /** 系统角色 ID */
  id: string;
  /** 系统角色标识 */
  key: string;
  /** 系统角色名称 */
  label: string;
  /** 角色说明 */
  description: string;
  /** 是否内置角色 */
  builtin: boolean;
  /** 获授该角色的用户数量 */
  userCount: number;
  /** 创建时间 */
  createdTime?: string;
  /** 更新时间 */
  updatedTime?: string;
}